backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
mediapipe==0.10.8
opencv-python==4.8.1.78
numpy==1.26.2
python-multipart==0.0.6
openai==1.3.5
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
websockets==12.0
pillow==10.1.0

# Pose comparison and DTW functionality
dtaidistance==2.3.10

# Audio processing for synchronized interface
pydub==0.25.1
simpleaudio==1.0.4
